Refactoring to simplify
