+++ lightgbm/src/main/scala/com/microsoft/ml/spark/lightgbm/LightGBMUtils.scala
@@ -6,20 +6,19 @@ package com.microsoft.ml.spark.lightgbm
 import java.io._
 import java.net.{ServerSocket, Socket}
 import java.util.concurrent.Executors
-
 import com.microsoft.ml.lightgbm._
 import com.microsoft.ml.spark.core.env.NativeLoader
 import com.microsoft.ml.spark.core.utils.ClusterUtil
 import com.microsoft.ml.spark.featurize.{Featurize, FeaturizeUtilities}
-import com.microsoft.ml.spark.lightgbm.dataset.LightGBMDataset
+import com.microsoft.ml.spark.lightgbm.dataset.{DenseAggregatedColumns, LightGBMDataset}
 import com.microsoft.ml.spark.lightgbm.params.TrainParams
 import org.apache.spark.{SparkEnv, TaskContext}
 import org.apache.spark.ml.PipelineModel
 import org.apache.spark.ml.attribute._
-import org.apache.spark.sql.{DataFrame, Dataset}
+import org.apache.spark.sql.{DataFrame, Dataset, SparkSession}
 import org.slf4j.Logger
-
 import ConnectionState._
+import com.microsoft.ml.spark.lightgbm.dataset.DatasetUtils.getSlotNames
 
 import scala.collection.immutable.HashSet
 import scala.collection.mutable.ListBuffer
@@ -55,49 +54,14 @@ object LightGBMUtils {
                     groupColumn: Option[String] = None,
                     oneHotEncodeCategoricals: Boolean = true): PipelineModel = {
     // Create pipeline model to featurize the dataset
-    val featuresToHashTo = FeaturizeUtilities.NumFeaturesTreeOrNNBased
     val featureColumns = dataset.columns.filter(col => col != labelColumn &&
       !weightColumn.contains(col) && !groupColumn.contains(col)).toSeq
-    val featurizer = new Featurize()
+    new Featurize()
       .setOutputCol(featuresColumn)
       .setInputCols(featureColumns.toArray)
       .setOneHotEncodeCategoricals(oneHotEncodeCategoricals)
-      .setNumFeatures(featuresToHashTo)
-    featurizer.fit(dataset)
-  }
-
-  def getCategoricalIndexes(df: DataFrame,
-                            featuresCol: String,
-                            slotNames: Array[String],
-                            categoricalColumnIndexes: Array[Int],
-                            categoricalColumnSlotNames: Array[String]): Array[Int] = {
-    val categoricalIndexes = if(slotNames.nonEmpty) {
-      categoricalColumnSlotNames.map(slotNames.indexOf(_))
-    } else {
-      val categoricalSlotNamesSet = HashSet(categoricalColumnSlotNames: _*)
-      val featuresSchema = df.schema(featuresCol)
-      val metadata = AttributeGroup.fromStructField(featuresSchema)
-      if (metadata.attributes.isEmpty) Array[Int]()
-      else {
-        metadata.attributes.get.zipWithIndex.flatMap {
-          case (null, _) => Iterator()
-          case (attr, idx) =>
-            if (attr.name.isDefined && categoricalSlotNamesSet.contains(attr.name.get)) {
-              Iterator(idx)
-            } else {
-              attr match {
-                case _: NumericAttribute | UnresolvedAttribute => Iterator()
-                // Note: it seems that BinaryAttribute is not considered categorical,
-                // since all OHE cols are marked with this, but StringIndexed are always Nominal
-                case _: BinaryAttribute => Iterator()
-                case _: NominalAttribute => Iterator(idx)
-              }
-            }
-        }
-      }
-    }
-
-    categoricalColumnIndexes.union(categoricalIndexes).distinct
+      .setNumFeatures(FeaturizeUtilities.NumFeaturesTreeOrNNBased)
+      .fit(dataset)
   }
 
   def sendDataToExecutors(hostAndPorts: ListBuffer[(Socket, String)], allConnections: String): Unit = {
@@ -147,59 +111,11 @@ object LightGBMUtils {
     }
   }
 
-  /**
-    * Opens a socket communications channel on the driver, starts a thread that
-    * waits for the host:port from the executors, and then sends back the
-    * information to the executors.
-    *
-    * @param numTasks The total number of training tasks to wait for.
-    * @return The address and port of the driver socket.
-    */
-  def createDriverNodesThread(numTasks: Int, df: DataFrame,
-                              log: Logger, timeout: Double,
-                              barrierExecutionMode: Boolean,
-                              driverServerPort: Int): (String, Int, Future[Unit]) = {
-    // Start a thread and open port to listen on
-    implicit val context: ExecutionContextExecutor =
-      ExecutionContext.fromExecutor(Executors.newSingleThreadExecutor())
-    val driverServerSocket = new ServerSocket(driverServerPort)
-    // Set timeout on socket
-    val duration = Duration(timeout, SECONDS)
-    if (duration.isFinite()) {
-      driverServerSocket.setSoTimeout(duration.toMillis.toInt)
-    }
-    val f = Future {
-      var emptyTaskCounter = 0
-      val hostAndPorts = ListBuffer[(Socket, String)]()
-      if (barrierExecutionMode) {
-        log.info(s"driver using barrier execution mode")
-        def connectToWorkers: Boolean = handleConnection(driverServerSocket, log,
-          hostAndPorts) == Finished || connectToWorkers
-        connectToWorkers
-      } else {
-        log.info(s"driver expecting $numTasks connections...")
-        while (hostAndPorts.size + emptyTaskCounter < numTasks) {
-          val connectionResult = handleConnection(driverServerSocket, log, hostAndPorts)
-          if (connectionResult == ConnectionState.EmptyTask) emptyTaskCounter += 1
-        }
-      }
-      // Concatenate with commas, eg: host1:port1,host2:port2, ... etc
-      val allConnections = hostAndPorts.map(_._2).mkString(",")
-      log.info(s"driver writing back to all connections: $allConnections")
-      // Send data back to all tasks and helper tasks on executors
-      sendDataToExecutors(hostAndPorts, allConnections)
-      closeConnections(log, hostAndPorts, driverServerSocket)
-    }
-    val host = ClusterUtil.getDriverHost(df)
-    val port = driverServerSocket.getLocalPort
-    log.info(s"driver waiting for connections on host: $host and port: $port")
-    (host, port, f)
-  }
 
   /** Returns an integer ID for the current worker.
     * @return In cluster, returns the executor id.  In local case, returns the partition id.
     */
-  def getWorkerId(): Int = {
+  def getWorkerId: Int = {
     val executorId = SparkEnv.get.executorId
     val ctx = TaskContext.get
     val partId = ctx.partitionId
@@ -212,7 +128,7 @@ object LightGBMUtils {
   /** Returns true if spark is run in local mode.
     * @return True if spark is run in local mode.
     */
-  def isLocalExecution(): Boolean = {
+  def isLocalExecution: Boolean = {
     val executorId = SparkEnv.get.executorId
     executorId == "driver"
   }
@@ -220,7 +136,7 @@ object LightGBMUtils {
   /** Returns a unique task Id for the current task run on the executor.
     * @return A unique task id.
     */
-  def getTaskId(): Long = {
+  def getTaskId: Long = {
     val ctx = TaskContext.get
     val taskId = ctx.taskAttemptId()
     taskId
@@ -250,28 +166,6 @@ object LightGBMUtils {
       (if (trainParams.categoricalFeatures.isEmpty) ""
       else s"categorical_feature=${trainParams.categoricalFeatures.mkString(",")}")
     datasetParams
-  }
-
-  def generateDenseDataset(numRows: Int, numCols: Int, featuresArray: SWIGTYPE_p_double,
-                           referenceDataset: Option[LightGBMDataset],
-                           featureNamesOpt: Option[Array[String]],
-                           trainParams: TrainParams, chunkSize: Int): LightGBMDataset = {
-    val isRowMajor = 1
-    val datasetOutPtr = lightgbmlib.voidpp_handle()
-    val datasetParams = getDatasetParams(trainParams)
-    val data64bitType = lightgbmlibConstants.C_API_DTYPE_FLOAT64
-    try {
-      // Generate the dataset for features
-      LightGBMUtils.validate(lightgbmlib.LGBM_DatasetCreateFromMat(lightgbmlib.double_to_voidp_ptr(featuresArray),
-        data64bitType, numRows, numCols,
-        isRowMajor, datasetParams, referenceDataset.map(_.datasetPtr).orNull, datasetOutPtr),
-        "Dataset create")
-    } finally {
-      lightgbmlib.delete_doubleArray(featuresArray)
-    }
-    val dataset = new LightGBMDataset(lightgbmlib.voidpp_value(datasetOutPtr))
-    dataset.setFeatureNames(featureNamesOpt, numCols)
-    dataset
   }
 
   /** Generates a sparse dataset in CSR format.
